move Shapelets class into classify module
